# Optimized version of i-am-closer's solution with pan's architecture