--- a/2022/04/pan93412-fast/Cargo.toml
+++ b/2022/04/pan93412-fast/Cargo.toml
@@ -12,8 +12,6 @@ codegen-units = 1
 
 [dependencies]
 anyhow = "1.0.66"
-delegate-attr = "0.2.9"
-hashbrown = { version = "0.13.1", features = ["nightly"] }
 
 [dev-dependencies]
 paste = "1.0.9"