--- a/2022/04/pan93412-fast/rust-toolchain.toml
+++ b/2022/04/pan93412-fast/rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+channel = "nightly"