--- a/2022/04/pan93412-fast/src/lib.rs
+++ b/2022/04/pan93412-fast/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod assignment;
+pub mod section;