--- a/2022/04/pan93412-fast/src/section.rs
+++ b/2022/04/pan93412-fast/src/section.rs
@@ -57,7 +57,7 @@ impl Section {
     #[inline]
     pub fn is_subset(&self, other: &Self) -> bool {
         // SELF         |--------------|
-        // THEIR        |------|-|-----|
+        // OTHER        |------|-|-----|
         self.lower() <= other.lower() && self.upper() >= other.upper()
     }