--- a/2022/04/pan93412-fast/tests/testdata
+++ b/2022/04/pan93412-fast/tests/testdata
@@ -1,0 +1,1 @@
+../../pan93412/tests/testdata