--- a/2022/04/pan93412/README.md
+++ b/2022/04/pan93412/README.md
@@ -1,6 +1,6 @@
 # Pan's solution of AoC 2022 Day 4
 
-The performance-oriented and well-engineered solution.
+The well-engineered and extensible solution.
 
 ## Run