… 2022/04/pan93412-fast/LICENSE
… 2022/04/pan93412-fast/docs/optimized-sol.webp